t2 method 2 edits
Changed the x-value calculation for t2 method 2 determination. Now, user
enters the delay time used, and this is used as the x-value for each
maximum found. Note: very important to edit the threshold for each
calculation - open excel and do a sample plot to confirm data is
appropriate; edit threshold accordingly.